fix(errors): Change type for code, for avoid negatives

+++ src/dankcompiler/errors/TokenError.java
@@ -1,13 +1,13 @@
 package dankcompiler.errors;
 
 public class TokenError {
-    public final int code;
+    public final long code;
     public final TokenErrorType type;
     public final String lexem;
     public final int line;
     public final int column;
     public final TokenErrorCode errorcode;
-    public TokenError(int code, TokenErrorType type, String lexem, int line, int column, TokenErrorCode errorcode){
+    public TokenError(long code, TokenErrorType type, String lexem, int line, int column, TokenErrorCode errorcode){
         this.code = code;
         this.type = type;
         this.lexem = lexem;
